logout-AuthUser-FileUpload

+++ enstruct/src/main/java/com/example/enstruct/controller/AssignmentController.java
@@ -7,14 +7,22 @@ import com.example.enstruct.service.IAttachmentService;
 import com.example.enstruct.service.ISubmissionService;
 import com.example.enstruct.service.IUserService;
 import com.example.enstruct.util.AuthManager;
+import org.apache.tomcat.jni.Local;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
 
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.text.DateFormat;
 import java.time.LocalDate;
+import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
@@ -38,6 +46,12 @@ public class AssignmentController {
 
     @GetMapping("/assignment-add")
     public String addAssignment(Model model) {
+
+        String authenticationResponse = AuthManager.getInstance().labelUser(true);
+        if(authenticationResponse != "continue") {
+            return authenticationResponse;
+        }
+
         model.addAttribute("assignment", new Assignment());
         model.addAttribute("courses", classesService.getClasses());
         return "addAssignment";
@@ -45,6 +59,12 @@ public class AssignmentController {
 
     @PostMapping("/assignment-add")
     public String addAssignmentSubmit(@ModelAttribute Assignment assignment, Model model) {
+
+        String authenticationResponse = AuthManager.getInstance().labelUser(true);
+        if(authenticationResponse != "continue") {
+            return authenticationResponse;
+        }
+
         model.addAttribute("assignment", assignment);
         Classes course = assignment.getCourse();
         assignmentService.addAssignment(assignment);
@@ -53,6 +73,14 @@ public class AssignmentController {
 
     @GetMapping("/assignment/{id}")
     public ModelAndView getAssignment(@PathVariable long id) {
+
+        String authenticationResponse = AuthManager.getInstance().labelUser(true);
+        if(authenticationResponse != "continue") {
+            ModelAndView mv = new ModelAndView();
+            mv.setViewName(authenticationResponse);
+            return mv;
+        }
+
         Assignment assignment = assignmentService.getAssignment(id);
         ModelAndView mv = new ModelAndView();
         mv.addObject("assignment", assignment);
@@ -63,6 +91,14 @@ public class AssignmentController {
 
     @GetMapping("/assignment/{id}/edit")
     public ModelAndView updateAssignment(@PathVariable long id) {
+
+        String authenticationResponse = AuthManager.getInstance().labelUser(true);
+        if(authenticationResponse != "continue") {
+            ModelAndView mv = new ModelAndView();
+            mv.setViewName(authenticationResponse);
+            return mv;
+        }
+
         Assignment assignment = assignmentService.getAssignment(id);
         ModelAndView mv = new ModelAndView();
         mv.addObject("assignment", assignment);
@@ -73,6 +109,14 @@ public class AssignmentController {
 
     @PostMapping("/assignment/{id}/edit")
     public ModelAndView updateAssignmentSubmit(@PathVariable long id, @ModelAttribute Assignment assignment, Model model) {
+
+        String authenticationResponse = AuthManager.getInstance().labelUser(true);
+        if(authenticationResponse != "continue") {
+            ModelAndView mv = new ModelAndView();
+            mv.setViewName(authenticationResponse);
+            return mv;
+        }
+
         model.addAttribute("assignment", assignment);
         Classes course = assignment.getCourse();
         assignmentService.addAssignment(assignment);
@@ -82,6 +126,14 @@ public class AssignmentController {
 
     @PostMapping("/assignment/{id}/delete")
     public ModelAndView deleteAssignment(@PathVariable long id) {
+
+        String authenticationResponse = AuthManager.getInstance().labelUser(true);
+        if(authenticationResponse != "continue") {
+            ModelAndView mv = new ModelAndView();
+            mv.setViewName(authenticationResponse);
+            return mv;
+        }
+
         Assignment assignment = assignmentService.getAssignment(id);
         Classes course = assignment.getCourse();
         assignmentService.deleteAssignment(id);
@@ -91,6 +143,14 @@ public class AssignmentController {
     // get all submissions in assignment
     @GetMapping("/assignment/{id}/submissions")
     public ModelAndView getAllSubmissionsInAssignment(@PathVariable long id) {
+
+        String authenticationResponse = AuthManager.getInstance().labelUser(true);
+        if(authenticationResponse != "continue") {
+            ModelAndView mv = new ModelAndView();
+            mv.setViewName(authenticationResponse);
+            return mv;
+        }
+
         List<Submission> submissions = submissionService.getAllSubmissionsInAssignment(id);
         ModelAndView mv = new ModelAndView();
         mv.addObject("submissions", submissions);
@@ -102,6 +162,14 @@ public class AssignmentController {
     // get submission of student in assignment
     @GetMapping("/assignment/{assignmentId}/{studentNumber}")
     public ModelAndView getSubmissionOfStudentInAssignment(@PathVariable long assignmentId, @PathVariable long studentNumber) {
+
+        String authenticationResponse = AuthManager.getInstance().labelUser(true);
+        if(authenticationResponse != "continue") {
+            ModelAndView mv = new ModelAndView();
+            mv.setViewName(authenticationResponse);
+            return mv;
+        }
+
         Submission submission = submissionService.getSubmissionOfStudentInAssignment(studentNumber, assignmentId);
         ModelAndView mv = new ModelAndView("showSubmission");
         mv.addObject("submission", submission);
@@ -113,24 +181,69 @@ public class AssignmentController {
     //STUDENT SIDE
     @RequestMapping(value = "/student/assignment/{assignmentId}", method = RequestMethod.GET)
     public String showStudentAssignment(Model model, @PathVariable long assignmentId) {
-        System.out.println("IN HERE");
+
+        String authenticationResponse = AuthManager.getInstance().labelUser(false);
+        if(authenticationResponse != "continue") {
+            return authenticationResponse;
+        }
+
+        AuthManager am = AuthManager.getInstance();
+        User u = am.getLoggedInUser();
+
+
         Assignment a = assignmentService.getAssignment(assignmentId);
+
+
+        //if nakapasa
+
+        if(submissionService.getSubmissionOfStudentInAssignment(u.getUserId(), assignmentId) == null) {
+            model.addAttribute("submitted", false);
+        } else {
+            model.addAttribute("submitted", true);
+        }
+
+
+        //if tapos na deadline
+        LocalDate dateNow = LocalDate.now();
+        LocalTime timeNow = LocalTime.now();
+
+        if(dateNow.isBefore(a.getDeadline_date())) {
+            model.addAttribute("available", true);
+        } else if(dateNow.isAfter(a.getDeadline_date())) {
+            model.addAttribute("available", false);
+        } else if(dateNow.isEqual(a.getDeadline_date())) {
+            if (timeNow.isBefore(a.getDeadline_time())) {
+                model.addAttribute("available", true);
+            } else {
+                model.addAttribute("available", false);
+            }
+        }
+
 
 
         model.addAttribute("instructions", a.getInstruction());
 
         LocalDate d = a.getDeadline_date();
-        DateFormat df = DateFormat.getDateInstance();
-        //String deadline = df.format(d);
-        //System.out.println("DEADLINE: " + deadline);
-        model.addAttribute("deadline", a.getDeadline_date());
+        LocalTime t = a.getDeadline_time();
+
+        String deadline = t.format(DateTimeFormatter.ofPattern("hh:mm a")) + ", " + d.format(DateTimeFormatter.ofPattern("MMMM dd, yyyy"));
+        model.addAttribute("assignmentName", a.getName());
+        model.addAttribute("deadline", deadline);
+
         model.addAttribute("maxScore", a.getMaxScore());
         model.addAttribute("assignmentId", assignmentId);
         return "studentAssignment";
     }
 
     @RequestMapping(value = "/student/assignment/{assignmentId}/submit/{filename}", method = RequestMethod.POST)
-    public ModelAndView showStudentAssignment(Model model, @PathVariable long assignmentId, @PathVariable String filename) {
+    public ModelAndView showStudentAssignment(Model model, @PathVariable long assignmentId, @PathVariable String filename, @RequestParam("upload") MultipartFile file) {
+
+        String fil = "C://enstruct//repository//";
+
+        String authenticationResponse = AuthManager.getInstance().labelUser(false);
+        if(authenticationResponse != "continue") {
+            return new ModelAndView(authenticationResponse, (Map<String, ?>) model);
+        }
 
         AuthManager am = AuthManager.getInstance();
         User user = am.getLoggedInUser();
@@ -145,7 +258,6 @@ public class AssignmentController {
 
         attachmentService.addAttachment(a);
 
-
         a = attachmentService.getInsertedAttachment(a.getFilename(), user.getUserId());
 
 
@@ -156,6 +268,20 @@ public class AssignmentController {
         s.setStudentNumber(u);
 
         submissionService.addSubmission(s);
+
+        try {
+
+            // Get the file and save it somewhere
+            byte[] bytes = file.getBytes();
+            fil += String.valueOf(assignmentId) + "//";
+            Files.createDirectories(Paths.get(fil));
+            Path path = Paths.get(fil + file.getOriginalFilename());
+            Files.write(path, bytes);
+
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
 
         s = submissionService.getLatestSubmission();
